refactor: split component_graph.rs into directory module
Break the monolithic 3,770-line component_graph.rs into a directory
module with focused sub-files:

- node.rs: ComponentNode, ComponentKind, ComponentStatusHandle,
  RelationshipKind, and update channel types
- graph.rs: ComponentGraph struct/impl, validation functions,
  snapshot types, and constants
- transaction.rs: GraphTransaction struct, impl, and Drop
- wait.rs: wait_for_status() free function
- tests.rs: all unit tests (~1,938 lines)
- mod.rs: module declarations and re-exports

All public imports remain unchanged. Internal cross-module access
uses pub(super) visibility for graph fields and helper methods.

Also includes minor formatting cleanup in mysql_helpers.rs.
